requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
deepagents>=0.1.0
pydantic>=2.0.0
jinja2>=3.1.0
python-multipart>=0.0.9
aiofiles>=24.1.0
markdown>=3.7
PyPDF2>=3.0.0
python-docx>=1.1.0
